### 2023-06-26 #5589 @editorialbot
[PRE REVIEW]: PATATO: a Python photoacoustic tomography analysis toolkit
**Submitting author:** @tomelse (<a href="http://orcid.org/0000-0002-2652-4190">Thomas Else</a>)
**Repository:** https://github.com/BohndiekLab/patato
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
