+++ .github/workflows/build_wheels.yml
@@ -4,8 +4,7 @@ on:
   push:
     branches:
       - main
-
-#on: [push, pull_request]
+      - develop
 
 jobs:
   build_wheels:
@@ -31,3 +30,16 @@ jobs:
       - uses: actions/upload-artifact@v3
         with:
           path: ./wheelhouse/*.whl
+
+  build_sdist:
+    name: Build source distribution
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Build sdist
+        run: pipx run build --sdist
+
+      - uses: actions/upload-artifact@v3
+        with:
+          path: dist/*.tar.gz
