+++ .github/workflows/draft-pdf.yml
@@ -5,10 +5,6 @@ on:
     branches:
       - develop
       - main
-  pull_request:
-    branches:
-      - develop
-      - main
 
 jobs:
   paper:
