### 2023-06-28 8431291
Add loading bar for model-based reconstruction
