+++ patato/recon/model_based/model_based.py
@@ -296,7 +296,7 @@ class ModelBasedReconstruction(ReconstructionAlgorithm):
         t = t.reshape((np.product(t.shape[:-2]),) + t.shape[-2:])
         output = np.zeros((t.shape[0],) + output_shape)
 
-        for i, t0 in tqdm(enumerate(t)):
+        for i, t0 in enumerate(tqdm(t)):
             result = model(t0.flatten()).reshape(output_shape)
             if gpu:
                 output[i] = result.get()
