### 2023-06-30 3cdbd3c
Remove old way of processing ultrasound data
