### 2023-07-05 #5589 @AoifeHughes
I'm having a bit of trouble finding reviewers; @tomelse would you have any suggestions either on the [JOSS reviewers]( https://reviewers.joss.theoj.org/reviewers ) table or people in your field who would be able to review?
