+++ patato/unmixing/learned_unmixer.py
@@ -1,0 +1,108 @@
+#  Copyright (c) Janek Grohl 2023.
+#  License: MIT
+
+from typing import Tuple, Optional, List, Union
+
+import numpy as np
+
+from learned_PA_sO2_estimator.utils.io import convert_numpy_array
+from learned_PA_sO2_estimator.estimate_sO2 import evaluate
+
+from ..core.image_structures.reconstruction_image import Reconstruction
+from ..core.image_structures.unmixed_image import UnmixedData
+from ..core.image_structures.single_parameter_data import SingleParameterData
+from ..io.attribute_tags import UnmixingAttributeTags
+from ..processing.processing_algorithm import SpatialProcessingAlgorithm
+from ..unmixing.spectra import Spectrum, SPECTRA_NAMES
+
+
+class LearnedSpectralUnmixer(SpatialProcessingAlgorithm):
+    """
+    The LearnedSpectralUnmixer class can be used to unmix multispectral photoacoustic data with a machine learning-based
+    unmixing algorithm that is trained on a specific training data set.
+
+    This takes in reconstruction data and spits out spectrally unmixed data.
+    """
+
+    @staticmethod
+    def re_grid(reconstruction: np.ndarray, scaling_factor: int) -> np.ndarray:
+        """
+        This is copy-pasted from the linear unmixer class.
+        There might be a good case for writing an abstract base class to avoid redundant code.
+
+        This rescaling applies a smoothing function by convolving the image with a kernel of ones.
+        Then it extracts N equidistant slices from the original image, where N is calculated from the scaling factor.
+
+        The blurring is introduced to prevent violating nyquist theorem by removing high frequencies from the image.
+
+        Parameters
+        ----------
+        reconstruction
+        scaling_factor
+
+        Returns
+        -------
+        a numpy array that is rescaled.
+
+        """
+        if scaling_factor == 1:
+            return reconstruction
+        else:
+            n_spatial = 3 if reconstruction.ndim == 5 else 2
+            reshape_kernel_size = tuple(min(scaling_factor, size) for size in reconstruction.shape[-n_spatial:])
+            extend = (None,) * len(reconstruction.shape[:-n_spatial])
+            kernel = np.ones(reshape_kernel_size)[extend]
+            kernel /= np.sum(kernel)
+            from scipy.signal import convolve
+            smoothed = convolve(reconstruction, kernel, mode="same")
+            slice_selection = (slice(None, None),) * len(reconstruction.shape[:-n_spatial])
+            slice_selection += (slice(None, None, scaling_factor),) * n_spatial
+            return smoothed[slice_selection]
+
+    def run(self, reconstruction: Reconstruction, _, **kwargs) -> Tuple[SingleParameterData, dict, None]:
+        # Select the right wavelengths:
+        wavelengths = self.wavelengths
+        wavelength_indices = np.where(np.isclose(reconstruction.wavelengths[:, None], wavelengths[None, :]))[0]
+
+        # Get the reconstructed data
+        recon_data = reconstruction[:, wavelength_indices].raw_data
+
+        # Change the grid
+        recon_data = self.re_grid(recon_data, self.rescaling_factor)
+
+        original_shape = np.shape(recon_data)
+        n_frames = original_shape[0]
+        n_wl = original_shape[1]
+
+        # reshape recon_data to (n_samples, n_wavelengths) from [frames, wavelengths, x, y, z]
+        recon_data = np.swapaxes(recon_data, 0, 1)
+        recon_data = np.reshape(recon_data, (n_wl, -1))
+        recon_data = np.swapaxes(recon_data, 0, 1)
+
+        # Unmix.
[…]
