+++ pyproject.toml
@@ -73,6 +73,7 @@ patato-import-ithera = "patato.convenience_scripts.convert_binary_to_hdf5:main"
 patato-draw-roi = "patato.convenience_scripts.draw_roi:main"
 patato-set-speed-of-sound = "patato.convenience_scripts.tune_speed_of_sound:main"
 patato-unmix = "patato.convenience_scripts.unmix:main"
+patato-learned_sO2 = "patato.convenience_scripts.learned_sO2:main"
 patato-reconstruct = "patato.convenience_scripts.process_msot:main"
 patato-analyse-dce = "patato.convenience_scripts.data_analysis.analyse_dce:main"
 patato-analyse-gc = "patato.convenience_scripts.data_analysis.analyse_gas_challenge:main"
