### 2023-07-19 65425e0
supress command line outputs and add attribute annotating the training dataset
