### 2023-07-20 ba0c822
Fix clinical detection, add roi options, allow single file speed of sound setting
