+++ .github/workflows/run_tests.yml
@@ -35,7 +35,6 @@ jobs:
         pip install coverage ruff
     - if: runner.os == 'Windows'
       run: |
-#        pip install jaxlib -f https://whls.blob.core.windows.net/unstable/index.html --use-deprecated legacy-resolver
         pip install jax[cpu]
     - name: Linting
       run: |
