### 2023-07-20 d1191ad
Update github workflow for Windows compatibility with JAX
