+++ .github/workflows/run_tests.yml
@@ -18,8 +18,11 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.8", "3.9", "3.10"]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
         os: [ubuntu-latest, windows-latest, macos-latest]
+        exclude:
+          - python-version: "3.8"
+            os: windows-latest
 
     steps:
     - uses: actions/checkout@v3
