+++ pyproject.toml
@@ -33,7 +33,7 @@ authors = [
 ]
 description = "A Python Toolkit for the analysis of photoacoustic tomography data."
 readme = "README.md"
-requires-python = ">=3.8,<3.11"
+requires-python = ">=3.8,<=3.11"
 
 keywords = [
     "photoacoustics",
