### 2023-07-20 ed41f4f
Update test special settings for windows- not needed.
