+++ .github/workflows/run_tests.yml
@@ -36,9 +36,6 @@ jobs:
         python -m pip install flake8 pytest
         pip install -e .
         pip install coverage ruff
-    - if: runner.os == 'Windows'
-      run: |
-        pip install jax[cpu]
     - name: Linting
       run: |
         ruff --format=github --select=E9,F63,F7,F82 --target-version=py37 . # Should do wider tests in the future.
