+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "patato"
-version = "0.5.3"
+version = "0.5.4"
 dependencies = [
     "requests",
     "cython",
