+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -1,0 +1,11 @@
+name: Release
+
+on:
+  release
+
+jobs:
+  build-and-publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: ./.github/workflows/run_tests.yml
+      - uses: ./.github/workflows/build_wheels.yml
