+++ .github/workflows/draft-pdf.yml
@@ -1,7 +1,7 @@
 name: Build draft PDF
 
 on:
-  push:
+  - push:
     branches:
       - develop
       - main
