+++ .github/workflows/build_wheels.yml
@@ -1,11 +1,11 @@
 name: Build wheels for PyPi
 
 on:
-  - push:
+  push:
     branches:
       - main
       - develop
-  - workflow_call
+  workflow_call:
 
 jobs:
   build_wheels:
