+++ .github/workflows/build_wheels.yml
@@ -6,6 +6,11 @@ on:
       - main
       - develop
   workflow_call:
+    - inputs:
+        name:
+          description: 'Display name and artifact name'
+          required: false
+          type: string
 
 jobs:
   build_wheels:
