+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -1,8 +1,8 @@
 name: Release
 
 on:
-  - release
-  - push
+  release:
+  push:
 
 jobs:
   test:
