+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -2,7 +2,11 @@ name: Release
 
 on:
   release:
+    types:
+      - created
   push:
+    branches:
+      - newworkflow
 
 jobs:
   test:
