+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -7,6 +7,8 @@ on:
   push:
     branches:
       - newworkflow
+  workflow_call:
+  workflow_dispatch:
 
 jobs:
   test:
