+++ .github/workflows/run_tests.yml
@@ -9,6 +9,7 @@ on:
     branches:
       - main
   workflow_call:
+  workflow_dispatch:
 
 
 env:
