+++ .github/workflows/build_wheels.yml
@@ -6,6 +6,7 @@ on:
       - main
       - develop
   workflow_call:
+  workflow_dispatch:
 
 jobs:
   build_wheels:
