+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -13,7 +13,7 @@ on:
 jobs:
   test:
     name: "Run tests"
-    uses: ./github/workflows/run_tests.yml
+    uses: BohndiekLab/patato/.github/workflows/run_tests.yml@main
   build_wheels:
     name: "Build wheels"
     uses: ./github/workflows/build_wheels.yml
