+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -13,7 +13,7 @@ on:
 jobs:
   test:
     name: "Run tests"
-    uses: BohndiekLab/patato/.github/workflows/run_tests.yml@main
+    uses: BohndiekLab/patato/.github/workflows/run_tests.yml@newworkflow
   build_wheels:
     name: "Build wheels"
     uses: ./github/workflows/build_wheels.yml
