### 2023-07-20 931cdcf
Make sure tests and building run consecutively
