+++ .github/workflows/draft-pdf.yml
@@ -2,10 +2,9 @@ name: Build draft PDF
 
 on:
   - push:
-    branches:
-      - develop
-      - main
-
+      branches:
+        - develop
+        - main
 jobs:
   paper:
     runs-on: ubuntu-latest
