+++ patato/_version.py
@@ -1,0 +1,4 @@
+# file generated by setuptools_scm
+# don't change, don't track in version control
+__version__ = version = '0.5.5.dev0+gde8e3c9.d20230720'
+__version_tuple__ = version_tuple = (0, 5, 5, 'dev0', 'gde8e3c9.d20230720')
