+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -11,12 +11,22 @@ on:
   workflow_dispatch:
 
 jobs:
+  test_run:
+    name: "HELP"
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Display version number
+        run: python -m setuptools_scm
+      - name: Show github tags
+        run: git tag | cat
   test:
     name: "Run tests"
     if: github.ref != 'refs/heads/newworkflow'
     uses: BohndiekLab/patato/.github/workflows/run_tests.yml@newworkflow
   build_wheels:
     name: "Build wheels"
+    if: github.ref != 'refs/heads/newworkflow'
     uses: BohndiekLab/patato/.github/workflows/build_wheels.yml@newworkflow
   publish:
     needs: build_wheels
