+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -18,6 +18,8 @@ jobs:
       - uses: actions/checkout@v3
         with:
           fetch-depth: 0
+      - name: Display structure of downloaded files
+        run: ls -R
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
         with:
@@ -25,10 +27,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install flake8 pytest
           pip install setuptools_scm
-          pip install -e .
-          pip install coverage ruff
       - name: Display version number
         run: python -m setuptools_scm
       - name: Show github tags
