+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -39,11 +39,10 @@ jobs:
     runs-on: ubuntu-latest
     name: Upload to PyPi.
     steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
       - uses: actions/download-artifact@v3
         with:
           path: dist
+      - name: Move artifacts to dist folder.
+        run: mv dist/artifact/* dist
       - name: Display structure of downloaded files
         run: ls -R
