+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -61,5 +61,7 @@ jobs:
           path: dist
       - name: Move artifacts to dist folder.
         run: mv dist/artifact/* dist
+      - name: Delete artifact folder.
+        run: rm dist/artifact
       - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
