+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -8,6 +8,7 @@ on:
     branches:
       - develop
       - main
+      - version_tracking
 
 jobs:
   test_run:
@@ -38,18 +39,28 @@ jobs:
     needs: build_wheels
     runs-on: ubuntu-latest
     name: Dummy upload to PyPI
+    environment:
+      name: testpypi
+      url: https://test.pypi.org/p/patato
     steps:
       - uses: actions/download-artifact@v3
         with:
           path: dist
       - name: Move artifacts to dist folder.
         run: mv dist/artifact/* dist
+      - name: Delete artifact folder.
+        run: rm dist/artifact
+      - name: Publish package distributions to TestPyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          repository-url: https://test.pypi.org/legacy/
       - name: Display structure of downloaded files
         run: ls -R
   publish:
     needs: build_wheels
     runs-on: ubuntu-latest
     name: Upload to PyPI.
+    if: github.event_name == 'release'
     environment:
       name: pypi
       url: https://pypi.org/p/patato
