+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -8,7 +8,6 @@ on:
     branches:
       - develop
       - main
-      - version_tracking
 
 jobs:
   test_run:
