+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -48,7 +48,7 @@ jobs:
       - name: Move artifacts to dist folder.
         run: mv dist/artifact/* dist
       - name: Delete artifact folder.
-        run: rm dist/artifact
+        run: rm -r dist/artifact
       - name: Publish package distributions to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
@@ -72,6 +72,6 @@ jobs:
       - name: Move artifacts to dist folder.
         run: mv dist/artifact/* dist
       - name: Delete artifact folder.
-        run: rm dist/artifact
+        run: rm -r dist/artifact
       - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
