+++ .github/workflows/upload_wheels_to_pypi.yml
@@ -8,6 +8,7 @@ on:
     branches:
       - develop
       - main
+      - version_tracking
 
 jobs:
   test_run:
