### 2023-07-21 79a2384
Fix workflow with no run outside of develop and main
