### 2023-07-21 dbf9fbc
Only upload to pypi on release
