### 2023-07-21 b96c26e
Relax jax requirements
