+++ pyproject.toml
@@ -22,7 +22,7 @@ dependencies = [
     "matplotlib_scalebar",
     "dask",
     "pybind11",
-    "jax[cpu]; platform_machine=='x86_64' or (platform_system=='Darwin' and platform_machine=='arm64')",
+    "jax[cpu]",
     # Need to verify that jax works on ARM platforms.
     "wget",
     "gdown"
