### 2023-07-24 #5686 @editorialbot
[REVIEW]: PATATO: a Python photoacoustic tomography analysis toolkit
**Submitting author:** @tomelse (<a href="http://orcid.org/0000-0002-2652-4190">Thomas Else</a>)
**Repository:** https://github.com/BohndiekLab/patato
**Branch with paper.md** (empty if default branch): main
**Version:** 0.5.13
