### 2023-07-26 #5686 @tomelse
I just noticed that the PDF above is a bit out of date, here is the updated version.
