### 2023-07-26 #5686 @tomelse
@AoifeHughes it looks like the paper PDF is an old version of the paper. Do you know why that would be? If you look at the current version, which is how I intended to submit it (https://github.com/BohndiekLab/patato/blob/main/paper/paper.md), the figures have been updated compared to the proof above. Would you be able to help fix this please?
