### 2023-07-26 9c1bc6b
Tidy up paper folder
