+++ paper/figure_code/paper.md
@@ -1,75 +1,0 @@
----
-title: 'PATATO: A Python Photoacoustic Analysis Toolkit'
-tags:
-  - Python, Photoacoustics, Imaging, Medical Imaging, Tomography, Optoacoustics, MSOT, Reconstruction, Backprojection
-authors:
-  - name: Thomas R. Else
-    orcid: 0000-0002-2652-4190
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: Janek Gröhl
-    affiliation: "1, 2"
-  - name: Lina Hacker
-    affiliation: "1, 2"
-  - name: Sarah E. Bohndiek
-    corresponding: true # (This is how to denote the corresponding author)
-    affiliation: "1, 2"
-affiliations:
-  - name:  CRUK Cambridge Institute, University of Cambridge, UK
-    index: 1
-  - name: Department of Physics, University of Cambridge, UK
-    index: 2
-date: 17th January 2023
-bibliography: paper.bib
----
-
-# Summary
-
-PATATO is a toolkit for processing and analyzing tomographic photoacoustic data. It offers fast implementations of commonly-used methods, including pre-processing, reconstruction, and time-series data analysis, via a convenient command-line interface or Python API. The toolkit utilizes [JAX](https://jax.readthedocs.io/) for GPU-accelerated pre-processing and backprojection, and [NumPy](https://numpy.org/doc/stable/index.html) for less intensive algorithms and easy integration with other pipelines. PATATO can take in data from many different sources, and outputs can be saved in HDF5 format for further analysis and easy transfer to other languages like R or MATLAB. The toolkit is open-source, hosted on [GitHub](https://github.com/tomelse/patato) and [PyPi](https://pypi.org/project/patato/), and welcomes contributions, bug-reports, and feedback. Detailed examples and an API reference is available at [https://patato.readthedocs.io/en/latest/](https://patato.readthedocs.io/en/latest/).
-
-# Statement of Need
-
-As a relatively new field, photoacoustic imaging (PAI) often lacks standardization in image processing methods. Despite the numerous algorithms published in literature for reconstruction, spectral analysis, and post-processing, access to the underlying code is limited, making it challenging to use the most current and effective methods [@Lutzweiler2013]. This lack of transparency may hinder reproducibility and slow the adoption of new methods within the community. While other tools for photoacoustic image reconstruction and analysis have been developed, they all lack open-source code for GPU acceleration and have limited documentation, making it difficult to extend their functionality [@Lafci2022; @OKelly2021]. PATATO offers an open-source solution with accelerated reconstruction algorithms, end-to-end analysis, an easy-to-use interface, open data format integration, and an extensible structure.
-
-Like many other imaging techniques, PAI requires multiple processing steps, many of which are computationally intensive. A typical pipeline for generating PA images includes filtering and interpolation of time-series data, delay-and-sum reconstruction, and linear spectral unmixing. The first two steps are particularly computationally intensive and require GPU implementations for efficient processing of large datasets, which can sometimes consist of thousands of frames. Additionally, the resulting images are highly sensitive to the parameters used, such as the cut-off frequency for filters, highlighting the need for efficient and transparent implementations of these algorithms to ensure consistent biological conclusions across datasets [@Shen2020]. In PATATO, these computationally intensive algorithms are implemented using JAX, a high-performance numerical computing toolkit, enabling GPU acceleration and consistent implementations between CPU and GPU [@jax2018github]. Additionally, PATATO stores all reconstruction parameters in the HDF5 file, making it straightforward to reproduce analyses.
-
-# Example Datasets
-
[…]
