+++ paper/figure_code/paper.bib
@@ -1,126 +1,0 @@
-@article{Tomaszewski2018,
-   abstract = {Measuring the functional status of tumor vasculature, including blood flow fluctuations and changes in oxygenation, is important in cancer staging and therapy monitoring. Current clinically approved imaging modalities suffer long procedure times and limited spatiotemporal resolution. Optoacoustic tomography (OT) is an emerging clinical imaging modality that may overcome these challenges. By acquiring data at multiple wavelengths,OTcan interrogate hemoglobin concentration and oxygenation directly and resolve contributions from injected contrast agents. In this study, we tested whether two dynamic OT techniques, oxygen-enhanced (OE) and dynamic contrast-enhanced (DCE)-OT, could provide surrogate biomarkers of tumor vascular function, hypoxia, and necrosis. We found that vascular maturity led to changes in vascular function that affected tumor perfusion, modulating the DCE-OT signal. Perfusion in turn regulated oxygen availability, driving the OE-OT signal. In particular, we demonstrate for the first time a strong per-tumor and spatial correlation between imaging biomarkers derived from these in vivo techniques and tumor hypoxia quantified ex vivo. Our findings indicate that OT may offer a significant advantage for localized imaging of tumor response to vascular-targeted therapies when compared with existing clinical DCE methods.},
-   author = {Michal R. Tomaszewski and Marcel Gehrung and James Joseph and Isabel Quiros Gonzalez and Jonathan A. Disselhorst and Sarah E. Bohndiek},
-   doi = {10.1158/0008-5472.CAN-18-1033},
-   issn = {0008-5472},
-   issue = {20},
-   journal = {Cancer Research},
-   month = {8},
-   pages = {canres.1033.2018},
-   publisher = {American Association for Cancer Research Inc.},
-   title = {Oxygen-enhanced and dynamic contrast-enhanced optoacoustic tomography provide surrogate biomarkers of tumour vascular function, hypoxia and necrosis.},
-   volume = {78},
-   url = {http://cancerres.aacrjournals.org/lookup/doi/10.1158/0008-5472.CAN-18-1033},
-   year = {2018},
-}
-
-@article{Hacker2021,
-   abstract = {Photoacoustic imaging (PAI) standardisation demands a stable, highly reproducible physical phantom to enable routine quality control and robust performance evaluation. To address this need, we have optimised a low-cost copolymer-in-oil tissue-mimickingmaterial formulation. The base material consists of mineral oil, copolymer and stabiliser with defined Chemical Abstract Service numbers. Speed of sound c(f) and acoustic attenuation coefficient α (f) were characterised over 2-10 MHz; optical absorption μ a ( λ ) and reduced scattering μ s '( λ ) coefficients over 450-900 nm. Acoustic properties were optimised by modifying base component ratios and optical properties were adjusted using additives. The temporal, thermomechanical and photo-stabilitywere studied, alongwith intra-laboratory fabrication and field-testing. c(f) could be tuned up to (1516±0.6) ms-1 and \α (f) to (17.4±0.3)dB cm -1 at 5 MHz. The base material exhibited negligible μ a ( λ ) and μ s '( λ ), which could be independently tuned by addition of Nigrosin or TiO2 respectively. These properties were stable over almost a year and were minimally affected by recasting. The material showed high intra-laboratory reproducibility (coefficient of variation <4% for c ( f), α (f), optical transmittance and reflectance), and good photo- and mechanical-stability in the relevant working range (20-40°C). The optimised copolymer-in-oil material represents an excellent candidate for widespread application in PAI phantoms, with properties suitable for broader use in biophotonics and ultrasound imaging standardisation efforts.},
[…]
