+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: bibliography.bib
 
 # Summary
 
-Photoacoustic imaging (PAI) is an emerging scalable imaging technology that combines the high contrast of optical imaging with the spatiotemporal resolution of ultrasound [@Beard2011]. Using light absorption by endogenous molecules, such as haemoglobin in red blood cells, PAI can reveal the emergence of diseases ranging from inflammation to cancer in both preclinical animal models and in patients [@Brown2019; @Regensburger2021; @Wang2012; @Steinberg2019]. Extracting accurate photoacoustic imaging biomarkers, such as blood oxygen saturation, from raw data requires a robust image reconstruction and analysis process, which is challenging due to the high dimensionality of the data across spatial, spectral and temporal domains. Here we introduce PATATO,  a Python toolkit that offers fast implementations of commonly-used data analysis methods, including pre-processing, reconstruction and temporal data analysis, via a user-friendly command-line interface and Python API. The toolkit uses [JAX](https://jax.readthedocs.io/), a modern machine learning tool, for GPU-accelerated pre-processing and image reconstruction, and [NumPy](https://numpy.org/doc/stable/index.html) for easy integration with other commonly used Python libraries.  PATATO is open-source, hosted on [GitHub](https://github.com/tomelse/patato) and [PyPi](https://pypi.org/project/patato/), and distributed under an MIT licence. We have designed PATATO to be modular and extendable to accommodate different data types, reconstruction methods, and custom analyses for specific scientific questions. We welcome contributions, bug reports, and feedback. Detailed examples, documentation, and an API reference are available at [https://patato.readthedocs.io/en/latest/](https://patato.readthedocs.io/en/latest/).
+Photoacoustic imaging (PAI) is an emerging scalable imaging technology that combines the high contrast of optical imaging with the spatiotemporal resolution of ultrasound [@Beard2011]. Using light absorption by endogenous molecules, such as haemoglobin in red blood cells, PAI can reveal the emergence of diseases ranging from inflammation to cancer in both preclinical animal models and in patients [@Brown2019; @Regensburger2021; @Wang2012; @Steinberg2019]. Extracting accurate photoacoustic imaging biomarkers, such as blood oxygen saturation, from raw data requires a robust image reconstruction and analysis process, which is challenging due to the high dimensionality of the data across spatial, spectral and temporal domains. Here we introduce PATATO,  a Python toolkit that offers fast implementations of commonly-used data analysis methods, including pre-processing, reconstruction and temporal data analysis, via a user-friendly command-line interface and Python API. The toolkit uses [JAX](https://jax.readthedocs.io/), a modern machine learning tool, for GPU-accelerated pre-processing and image reconstruction, and [NumPy](https://numpy.org/doc/stable/index.html) for easy integration with other commonly used Python libraries.  PATATO is open-source, hosted on [GitHub](https://github.com/BohndiekLab/patato) and [PyPi](https://pypi.org/project/patato/), and distributed under an MIT licence. We have designed PATATO to be modular and extendable to accommodate different data types, reconstruction methods, and custom analyses for specific scientific questions. We welcome contributions, bug reports, and feedback. Detailed examples, documentation, and an API reference are available at [https://patato.readthedocs.io/en/latest/](https://patato.readthedocs.io/en/latest/).
 
 # Statement of Need
 
