### 2023-08-03 a23ac95
Allow no overall correction factor
