### 2023-08-03 b740fff
Update overall correction factor None support
