+++ patato/__init__.py
@@ -62,3 +62,5 @@ for e in __all_exports:
     e.__module__ = __name__
 
 __all__ = [e.__name__ for e in __all_exports] + ["core", "io", "recon", "unmixing", "processing", "utils", "data"]
+
+print("HELLO")
