+++ patato/recon/backprojection_reference.py
@@ -44,7 +44,7 @@ class ReferenceBackprojection(ReconstructionAlgorithm):
         -------
 
         """
-
+        print(f"c = {speed_of_sound}")
         # Get parameters:
         dl = speed_of_sound / fs
 
