+++ patato/__init__.py
@@ -63,4 +63,3 @@ for e in __all_exports:
 
 __all__ = [e.__name__ for e in __all_exports] + ["core", "io", "recon", "unmixing", "processing", "utils", "data"]
 
-print("HELLO")
