### 2023-08-08 31701c7
Update image_sequence.py
