### 2023-08-09 4698cc4
allow for very small numbers instead of just zero
Added an upper limit of eps = 1e-10 to allow some lee way instead of one of the axes needing to be exactly 0.
