### 2023-08-09 3e60125
removed extra print statements
