### 2023-08-10 0c5d588
allow for a column to be constant, instead of close to zero
