### 2023-08-21 f04808c
Optionally do not run batches for some reconstruction algorithms
