+++ patato/recon/reconstruction_algorithm.py
@@ -26,6 +26,7 @@ class ReconstructionAlgorithm(TimeSeriesProcessingAlgorithm, ABC):
         self.n_pixels = n_pixels
         self.field_of_view = field_of_view
         self.custom_params = kwargs
+        self._batch = True
 
     @abstractmethod
     def reconstruct(self, raw_data: np.ndarray,
@@ -70,7 +71,7 @@ class ReconstructionAlgorithm(TimeSeriesProcessingAlgorithm, ABC):
             fov = self.field_of_view
         
         # Process in batches to avoid GPU running out of memory.
-        if time_series.shape[0] * time_series.shape[1] > PAT_MAXIMUM_BATCH_SIZE != -1:
+        if time_series.shape[0] * time_series.shape[1] > PAT_MAXIMUM_BATCH_SIZE != -1 and self._batch:
             new_recons = []
             ts_raw = time_series.raw_data
             shape = ts_raw.shape
