### 2023-08-21 aa0c747
New JAX Model based approach
