+++ patato/recon/jax_model_based/__init__.py
@@ -17,28 +17,33 @@ class JAXModelBasedReconstruction(ReconstructionAlgorithm):
     def __init__(self, n_pixels, field_of_view, pa_example: "PAData" = None, **kwargs
                  ):
         if pa_example is not None:
-            kwargs["geometry"] = kwargs.get("geometry", pa_example.get_scan_geometry())
-            kwargs["fs"] = kwargs.get("fs", pa_example.get_sampling_frequency())
-            kwargs["nt"] = kwargs.get("nt", pa_example.get_time_series().shape[-1])
-            kwargs["c"] = kwargs.get("c", pa_example.get_speed_of_sound())
-            kwargs["irf"] = kwargs.get("irf", pa_example.get_impulse_response())
+            kwargs["model_geometry"] = kwargs.get("model_geometry", pa_example.get_scan_geometry())
+            kwargs["model_fs"] = kwargs.get("model_fs", pa_example.get_sampling_frequency())
+            kwargs["model_nt"] = kwargs.get("model_nt", pa_example.get_time_series().shape[-1])
+            kwargs["model_c"] = kwargs.get("model_c", pa_example.get_speed_of_sound())
+            kwargs["model_irf"] = kwargs.get("model_irf", pa_example.get_impulse_response())
         super().__init__(n_pixels, field_of_view, **kwargs)
-
+        self._batch = False
         axes = [i for i, x in enumerate(n_pixels) if x != 1]
         dx = field_of_view[axes[0]] / (n_pixels[axes[0]] - 1)
         if dx != field_of_view[axes[1]] / (n_pixels[axes[1]] - 1) or field_of_view[axes[1]] != field_of_view[axes[0]]:
             raise ValueError("Current model based implementation only supports square reconstruction grids.")
 
-        detx = kwargs["geometry"][:, axes[0]]
-        dety = kwargs["geometry"][:, axes[1]]
-        fs = kwargs["fs"]
+        detx = kwargs["model_geometry"][:, axes[0]]
+        dety = kwargs["model_geometry"][:, axes[1]]
+        fs = kwargs["model_fs"]
         nx = n_pixels[axes[0]]
         x_0 = - field_of_view[axes[0]] / 2
-        nt = kwargs["nt"]
-        c = kwargs["c"]
+        nt = kwargs["model_nt"]
+        c = kwargs["model_c"]
         self._nx_model = nx
         self._model_matrix = self._generate_model(detx, dety, fs, dx, nx, x_0,
                                                   nt, c)
+        self._model_matrix.eliminate_zeros()
+        self._model_matrix.sort_indices()
+        from jax.experimental.sparse import CSR
+        self._model_matrix = CSR((self._model_matrix.data, self._model_matrix.indices, self._model_matrix.indptr), shape=self._model_matrix.shape)
+
 
     def _generate_model(self, detx: npt.ArrayLike, dety: npt.ArrayLike,
                         fs: float, dx: float, nx: int, x_0: float, nt: int,
@@ -65,37 +70,47 @@ class JAXModelBasedReconstruction(ReconstructionAlgorithm):
         model_matrix = get_model(detx, dety, dl, dx, nx, x_0, nt, cache=False)
         return model_matrix
 
+
     def reconstruct(self, raw_data: np.ndarray,
                     fs: float = None,
                     geometry: np.ndarray = None,
-                    n_pixels: Sequence[int] = None,
-                    field_of_view: Sequence[float] = None,
+                    n_pixels = None,
+                    field_of_view = None,
                     speed_of_sound=None,
                     **kwargs) -> np.ndarray:
         import jax
         import jax.numpy as jnp
+        from jax.experimental.sparse import CSR
         import jaxopt
[…]
