### 2023-08-22 d6c9146
Add regulariser to JAX model based implementation
