### 2023-08-22 2dfcecf
Update regulariser implementation for jax
