### 2023-08-22 a858e98
Static argnums for recon
