### 2023-08-22 c890276
Refactor way of doing regularisation in JAX model based
