### 2023-08-22 822a335
Remove partial static args
