### 2023-08-22 e9dae4b
Store most recent state
