### 2023-08-22 d0494d6
Speed up roi application
