### 2023-08-22 4181ba5
Hide tqdm after completion
