### 2023-08-25 9bb8544
Allow different constraints in optimisation :
