### 2023-09-17 BohndiekLab/patato#26 @MengjieSHI
errors when running tests 
Hi @tomelse,

I tested the option 1: using pip to install PATATO and tried to run tests. It seems the 'tests' folder is not downloaded, so I manually added it. The test failed because of three errors. I attached the screenshot of the error below.

![image](https://github.com/BohndiekLab/patato/assets/59768848/bebff889-eb61-4fea-9aaa-0489f4ca5f68)

The last two errors seem caused by the missing files due to the default relative path. Any suggestions to fix these?
