### 2023-09-17 #5686 @MengjieSHI
Hi @tomelse I created an [issue](https://github.com/BohndiekLab/patato/issues/26#issue-1899864930) regarding the errors when running the tests. Thanks!
