### 2023-09-18 BohndiekLab/patato#26 @tomelse
Hi @MengjieSHI,
Thanks so much for taking the time to review the code.

The layout of the package means that the tests have to be run on a locally installed version of PATATO. The way I've done this usually is the following:

```
git clone https://github.com/BohndiekLab/patato
cd patato
pip install -e .
cd tests
python -m unittest
```

This means that we don't have to ship all of the test code and example data with the installable version, reducing the amount of download time required for installing PATATO.

I hope that's okay? I tested it on Google Colab, and it seems to work.

Thanks,
Tom
