### 2023-09-18 aa7cf50
Update tests running documentation
