### 2023-09-18 BohndiekLab/patato#26 @MengjieSHI
Thanks for your fast response! Tests were running without issues now.

A minor comment on the [documentation](https://patato.readthedocs.io/en/latest/intro/03b_example_data.html) page, I guess you forgot to update the file names included in the example. I replaced them with'invivo_oe.hdf5' and 'invivo_dce.hdf5' and it works perfectly.
