### 2023-09-18 BohndiekLab/patato#29 @MengjieSHI
add clear contributing guidelines 
Hi @tomelse,

I saw there are examples about how to contribute new reconstruction algorithms and scan reader. It would be efficient to have a contribution page. I found [this](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) that could be helpful.

Best,
Mengjie

This issue is part of the JOSS [review](https://github.com/openjournals/joss-reviews/issues/5686)
