### 2023-09-19 BohndiekLab/patato#34 @MengjieSHI
Minor comments 
Hi @tomelse,  I am running the examples, and they work perfectly:)  Here are some minor comments I have for further improvements.

1. I guess the folder named study contains the folder named Scan_1, Scan_2. It would be great to rephrase the explanations a bit, as the files were also named Scan_1.

`patato-import-ithera /path/to/ithera/study/folder /path/to/data/folder`

2. For the GUI of tuning the speed-of-sound,
a. The captions of the subfigures were missing. Did the right figure show the profile along x or z direction?
b. Is it possible to reopen the GUI using the same command after I saved the speed-of-sound and quit it?  I failed on this, but I am wondering if there is a need to modify the speed-of-sound afterwards.

3. I am not very clear about how to rename the ROIs on the GUI. I clicked the rename button, but it seems it didn't work as expected.

4. A link is missing at line 3, first para, Examples 4. Figures got overlapping in the results of Block [10].

Thanks!

This is part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/5686#issuecomment-1726350602)
