### 2023-09-26 aa5a9c1
Update region drawing interface
