+++ patato/convenience_scripts/draw_roi.py
@@ -1,496 +1,420 @@
-#  Copyright (c) Thomas Else 2023.
-#  License: MIT
+from pathlib import Path
 
-"""
-``patato-draw-roi``: A script to draw regions of interest interactively on datasets..
-"""
-
-import argparse
-import glob
-from collections import namedtuple
-from os.path import join, split
-
-import matplotlib
 import matplotlib.pyplot as plt
+
+import tkinter as tk
+from tkinter import filedialog, Listbox, messagebox
+import customtkinter as ctk
 import numpy as np
-from matplotlib.widgets import Slider, Button, RangeSlider, PolygonSelector
-from .. import ImageSequence, ROI, PAData
-from ..utils import sort_key
-from scipy.optimize import least_squares
-
-if matplotlib.get_backend() == "MacOSX":
-    matplotlib.use("TkAgg")
-
+from matplotlib.widgets import PolygonSelector
+import time
+from .. import PAData, ROI
 from ..utils.roi_operations import ROI_NAMES, REGION_COLOURS, close_loop
 
-# TODO: Temporary fix for ?
 import os
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
 
-os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"
+from importlib.resources import files
 
-
-def circ_loss(x, points):
-    R = x[0]
-    x_0 = x[1]
-    y_0 = x[2]
-    return np.sum(((points[:, 0] - x_0) ** 2 / R ** 2 + (points[:, 1] - y_0) ** 2 / R ** 2 - 1) ** 2)
+POSITIONS = ["", "left", "right", "full", "radial", "ulnar", "superficial"]
 
 
-def get_circle_points(r, x_0, y_0, n_points=50):
-    thetas = np.linspace(0, 2 * np.pi, n_points, False)
-    return np.array([np.sin(thetas) * r + x_0, np.cos(thetas) * r + y_0]).T
+def shorten(text, width, placeholder=""):
+    if len(text) > width:
+        return text[:width] + placeholder
+    else:
+        return text
 
 
-def fitcirc(points):
-    points = np.array(points)
-    x_0 = np.mean(points[:, 0])
-    y_0 = np.mean(points[:, 1])
-    r_0 = np.sqrt(np.mean((points[:, 0] - x_0) ** 2 + (points[:, 1] - y_0) ** 2))
-    fit = least_squares(circ_loss, x0=np.array([r_0, x_0, y_0]), args=(points,))
-    return get_circle_points(*fit.x)
+class HDF5ViewerApp:
+    def __init__(self, root):
+        self.new_roi_vertices = None
+        self.pa_data_selected = None
+        self.polygon_selector = None
+        self.recon_map = {}
 
+        self.root = root
+        self.root.title("PATATO: Draw Region of Interest")
 
-positions = ["left", "right", "full", "", "radial", "ulnar", "superficial"]
+        # Set grid weights for columns and rows to allow resizing
+        self.root.columnconfigure(0, weight=1)
+        self.root.columnconfigure(1, weight=2)
+        self.root.columnconfigure(2, weight=1)
+        self.root.rowconfigure(0, weight=0)
+        self.root.rowconfigure(1, weight=0)
+        self.root.rowconfigure(2, weight=1)
 
+        self.create_widgets()  # Create the initial widgets
 
-def init_argparse():
-    parser = argparse.ArgumentParser(description="Process MSOT Data.")
-    parser.add_argument('input', type=str, help="Data Folder")
-    parser.add_argument('-n', '--name', type=str, help="ROI Name",
-                        choices=ROI_NAMES, default="unnamed")
[…]
