+++ pyproject.toml
@@ -25,7 +25,8 @@ dependencies = [
     "jax[cpu]",
     # Need to verify that jax works on ARM platforms.
     "wget",
-    "gdown"
+    "gdown",
+    "customtkinter"
 ]
 
 authors = [
@@ -70,7 +71,6 @@ namespaces = true  # true by default
 
 [project.scripts]
 patato-import-ithera = "patato.convenience_scripts.convert_binary_to_hdf5:main"
-patato-draw-roi = "patato.convenience_scripts.draw_roi:main"
 patato-set-speed-of-sound = "patato.convenience_scripts.tune_speed_of_sound:main"
 patato-unmix = "patato.convenience_scripts.unmix:main"
 patato-reconstruct = "patato.convenience_scripts.process_msot:main"
@@ -78,8 +78,11 @@ patato-analyse-dce = "patato.convenience_scripts.data_analysis.analyse_dce:main"
 patato-analyse-gc = "patato.convenience_scripts.data_analysis.analyse_gas_challenge:main"
 patato-convert-simpa = "patato.convenience_scripts.convert_simpa:main"
 
+[project.gui-scripts]
+patato-draw-roi = "patato.convenience_scripts.draw_roi:main"
+
 [tool.setuptools.package-data]
-"*" = ["*.cu", "*.cl", "*.json", "*.csv", "*.txt"]
+"*" = ["*.cu", "*.cl", "*.json", "*.csv", "*.txt", "*.png"]
 
 [tool.setuptools.exclude-package-data]
 "*" = [".gitattributes", ".gitignore"]
