### 2023-09-26 36bcac5
Add rename ROI functionality
