### 2023-09-26 5252a23
Fix rename Roi functionality
