### 2023-09-26 52b5505
Update ROI drawing interface
