### 2023-09-26 e5fffce
Sort draw roi hdf5s by date scanned
