### 2023-09-26 dda5d7c
Update README logo and speed of sound setting software.
