+++ patato/convenience_scripts/tune_speed_of_sound.py
@@ -1,222 +1,303 @@
-#  Copyright (c) Thomas Else 2023.
-#  License: MIT
-"""
-``patato-set-speed-of-sound``: This module contains functions for setting the speed of sound for all datasets in a
-directory.
-"""
-
-import argparse
 import glob
 import json
-import os
+from pathlib import Path
 
 import matplotlib.pyplot as plt
-import numpy as np
-from matplotlib.widgets import Slider, Button, RangeSlider
 
-from patato.recon import get_default_recon_preset
+import tkinter as tk
+from tkinter import filedialog, Listbox, messagebox
+import customtkinter as ctk
+
 from .. import PAData
+from ..recon import get_default_recon_preset
+from ..io.json.json_reading import read_reconstruction_preset
 
-try:
-    from pyopencl.array import Array
-except ImportError:
-    Array = None
+import os
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
 
-from .. import read_reconstruction_preset, PreProcessor,  OpenCLBackprojection
-from ..recon.reconstruction_algorithm import ReconstructionAlgorithm
-from ..utils import sort_key
+from importlib.resources import files
+
+
+def shorten(text, width, placeholder=""):
+    if len(text) > width:
+        return text[:width] + placeholder
+    else:
+        return text
+
+
+class HDF5ViewerApp:
+    def update_speed_of_sound(self, x):
+        c = self.slider3.get()
+        self.speed_label.configure(text=f"{c} m/s")
+        self.update_image()
+
+    def save_speed_of_sound(self):
+        self.pa_data_selected.set_speed_of_sound(self.slider3.get())
+
+    def __init__(self, root, start_file=None):
+        self.new_roi_vertices = None
+        self.pa_data_selected = None
+        self.polygon_selector = None
+        self.recon_map = {}
+
+        self.root = root
+        self.root.title("PATATO: Set Speed of Sound")
+
+        # Set grid weights for columns and rows to allow resizing
+        self.root.columnconfigure(0, weight=1)
+        self.root.columnconfigure(1, weight=2)
+        self.root.columnconfigure(2, weight=1)
+        self.root.rowconfigure(0, weight=0)
+        self.root.rowconfigure(1, weight=0)
+        self.root.rowconfigure(2, weight=1)
+
+        self.create_widgets()  # Create the initial widgets
+
+        if start_file is not None:
+            self.load_hdf5_folder(start_file)
+
+    def create_widgets(self):
+        # Create a label to display the loaded HDF5 file path
+        self.file_label = ctk.CTkLabel(self.root, text="No file loaded.")
+        self.file_label.grid(row=0, column=0, columnspan=3, pady=5)
+
+        # Create a button to load a folder containing HDF5 files
+        self.load_button = ctk.CTkButton(self.root, text="Load HDF5 Folder", command=self.load_hdf5_folder)
+        self.load_button.grid(row=1, column=0, columnspan=3, pady=5)
+
+        # Create a frame for additional widgets on the left
+        self.left_frame = ctk.CTkFrame(self.root, width=125)
+        self.left_frame.grid(row=2, column=0, padx=5, pady=5, sticky="nsew")
+
+        # Create a title label for the Listbox
+        self.listbox_title_label = ctk.CTkLabel(self.left_frame, text="HDF5 Files")
+        self.listbox_title_label.pack()
+
+        # Create a listbox to display the list of HDF5 files
+        self.file_listbox = Listbox(self.left_frame)
+        self.file_listbox.pack(fill=tk.BOTH, expand=True)
+        self.file_listbox.bind('<<ListboxSelect>>', self.load_selected_hdf5_file)
+
+        # Create a frame for Matplotlib and navigation toolbar in the center
[…]
