### 2023-09-26 f1266f8
Allow saving speed of sound to all
