### 2023-09-27 BohndiekLab/patato#34 @tomelse
Glad to hear the examples run perfectly! And apologies for the delayed response.

Could you elaborate a bit what you mean for part 1? Do you mean clarify in the Quick start guide in the documentation?

I've updated both the GUIs for ROI drawing and speed of sound tuning. I've wanted to update them for a long time... I've removed the line plot functionality as I don't think it's that helpful. Feel free to try it out and let me know if there's anything you think I've missed.

The same commands can still be used, but with no arguments this time.

```bash
patato-set-speed-of-sound
patato-draw-roi
```

I will work on number 4 today.

Thanks very much!
