### 2023-09-27 ab5ca6f
Add model-based attributes to saved reconstruction
