### 2023-10-02 BohndiekLab/patato#34 @tomelse
@MengjieSHI I've now fixed all of the issues you mentioned (see here: [https://patato.readthedocs.io/en/latest/examples/04_timeseriesanalysis.html](https://patato.readthedocs.io/en/latest/examples/04_timeseriesanalysis.html)).
