### 2023-10-03 BohndiekLab/patato#34 @MengjieSHI
Hi @tomelse Thanks for your updates! Yes, I mean it would be better to clarify a bit for Part 1 as the folder and files have the same name in your example.

I downloaded the latest release 0.5.12 from GitHub but seems the updates on GUIs were not applied, either for the commands. It is quite wired, anything you think I could try? Thanks!
