### 2023-10-03 BohndiekLab/patato#34 @tomelse
Have you run

`pip install --upgrade patato`

or `pip install --upgrade -e .`

to install the updated version?
