### 2023-10-03 BohndiekLab/patato#34 @MengjieSHI
Yes, just restarted my computer and everything works well now:)

The new GUIs look really great!

It seems there is no return after I clicked save speed of sound? I also got some warnings after closing the GUI.
![image](https://github.com/BohndiekLab/patato/assets/59768848/9ec46247-98d5-46d3-b0a4-e84c4802b811)
