### 2023-10-30 BohndiekLab/patato#40 @ASK-DataScience
Speed of sound using command line
Hi @tomelse,

I tried to run the command: patato-set-speed-of-sound

It showed the GUI and the example data and I could change the value. However the output shows this:

![image](https://github.com/BohndiekLab/patato/assets/7422503/c5c31a4b-e076-4b27-82e8-d0cb5589d339)
