### 2023-10-30 #5686 @ASK-DataScience
Hi @tomelse, I created an [issue](https://github.com/BohndiekLab/patato/issues/40) regarding running the speed of sound command on the command line. Thanks!
