### 2023-10-30 BohndiekLab/patato#40 @tomelse
Does the GUI work as expected? I think the errors/warnings come from tkinter, the toolkit used for making the GUI and shouldn't affect the functioning of the program.
