### 2023-10-30 BohndiekLab/patato#41 @ASK-DataScience
Errors while running tests
I am getting an error after running this step: pip install -e .

![image](https://github.com/BohndiekLab/patato/assets/7422503/96b1991b-037f-42db-9835-1e53dd095dd3)

Also after running this step: python -m unittest

![image](https://github.com/BohndiekLab/patato/assets/7422503/fa807f61-24e1-442e-b6e2-af523078da8c)

This is part of JOSS [Review](https://github.com/openjournals/joss-reviews/issues/5686)
