### 2023-10-30 #5686 @ASK-DataScience
Hi @tomelse, I created another [issue ](https://github.com/BohndiekLab/patato/issues/41)regarding the unit tests. Thanks!
