### 2023-10-30 BohndiekLab/patato#40 @ASK-DataScience
Yes, the GUI works as expected.
