### 2023-11-03 BohndiekLab/patato#41 @tomelse
The error with editable install is because Python requires the bundled C++ code to be compiled using Microsoft c++ build tools.

My intention for the unittests was to run the tests as follows. That way you don't need the compiler. I haven't worked out a good way of distributing the unittests through pip, so this seemed like the most sensible way to do it, given that most people won't be running unittests (unless you have any other suggestions?).

```
pip install patato
git clone https://github.com/BohndiekLab/patato
cd patato/tests
python -m unittest
```
