### 2023-11-06 BohndiekLab/patato#41 @tomelse
@ASK-DataScience do you think this is an acceptable way to do this?
