### 2023-11-06 BohndiekLab/patato#41 @ASK-DataScience
@tomelse I agree with you that most people won't be running unittests, but as a reviewer I should check if it is working correctly for me. Any suggestion on how to make it work? I followed your steps in the documentation.
