### 2023-11-07 BohndiekLab/patato#41 @tomelse
Did you try the instructions in my previous comment? I.e. installing but not editable (pip install patato), then separately git checkout the repo and run the tests on that code
