### 2023-11-07 BohndiekLab/patato#41 @tomelse
Otherwise you'd need to install the Microsoft C++ tools to compile the internal C++ code used for one of the algorithms
