### 2023-11-09 BohndiekLab/patato#41 @ASK-DataScience
@tomelse I just tried the instructions in your previous comment (not editable) and I got the following error:

![image](https://github.com/BohndiekLab/patato/assets/7422503/3410a992-3f62-4eed-abd1-8f456f221cec)
