### 2023-11-09 BohndiekLab/patato#41 @ASK-DataScience
So looks like the first error is resolved, but we are still stuck with the second one.
