### 2023-11-09 BohndiekLab/patato#41 @ASK-DataScience
patato is imported but still cannot run unittest

![image](https://github.com/BohndiekLab/patato/assets/7422503/bb889729-256c-4f5b-b0cf-969ae451acc2)
