### 2023-11-09 BohndiekLab/patato#41 @ASK-DataScience
OK. Tests are finally working for me. But I had to install and run patato on the "base" environment instead of its own environment.

![image](https://github.com/BohndiekLab/patato/assets/7422503/b2e61a70-1c4f-4a33-9ef8-180ed4dfbd5b)
