### 2023-11-09 BohndiekLab/patato#41 @ASK-DataScience
I suggest you add the option to run patato as not editable on the documentation page as well.
