### 2023-11-09 BohndiekLab/patato#41 @tomelse
Thanks, I'll add that to the documentation.
