### 2023-12-04 826a4d5
Update pyproject.toml to fix PyLops requirements
