### 2023-12-15 35765ec
Patch support for iThera 512 Echo
